Add tabIndex=-1 to links and search input
When a new tab is opened the focus is usually on the address bar, so not
inside the page. That means the event listener in main.js is ignored at
first until the page gets the focus, and the tab key causes focus to
jump directly to the first tabbable element. tabIndex="-1" prevents the
element from getting focus from keyboard navigation. That way when the
tab key is pressed in the address bar the focus is on the page, and not
on a tabbable element since none exist.